+++ src/ability/mod.rs
@@ -53,7 +53,6 @@ pub fn parse_suc_code(suc: &str) -> Vec<i32> {
 }
 
 pub use crate::ability::{
-    ai::ai_task::AiTaskExecutor as GxAIFun,
     assert::GxAssert,
     cmd::GxCmd,
     echo::GxEcho,