+++ src/loader.rs
@@ -16,17 +16,17 @@ use crate::err::*;
 
 use crate::ability::version::Version;
 use once_cell::sync::OnceCell;
+use orion_accessor::addr::Address;
+use orion_accessor::addr::GitRepository;
+use orion_accessor::addr::LocalPath;
+use orion_accessor::types::ResourceDownloader;
+use orion_accessor::update::DownloadOptions;
+use orion_accessor::update::UpdateScope;
 use orion_error::ContextRecord;
 use orion_error::ErrorConv;
 use orion_error::ErrorOwe;
 use orion_error::ErrorWith;
 use orion_error::WithContext;
-use orion_variate::addr::Address;
-use orion_variate::addr::GitRepository;
-use orion_variate::addr::LocalPath;
-use orion_variate::types::ResourceDownloader;
-use orion_variate::update::DownloadOptions;
-use orion_variate::update::UpdateScope;
 use orion_variate::vars::EnvDict;
 use orion_variate::vars::ValueDict;