+++ src/util/accessor.rs
@@ -1,12 +1,10 @@
 use crate::const_val::gxl_const::NET_ACCS_CTRL_PATH_FILE;
-use orion_conf::Yamlable;
-use orion_variate::{
-    addr::{
-        access_ctrl::serv::NetAccessCtrl,
-        accessor::{UniversalAccessor, UniversalConfig},
-    },
-    vars::{EnvDict, EnvEvalable},
+use orion_accessor::addr::{
+    access_ctrl::serv::NetAccessCtrl,
+    accessor::{UniversalAccessor, UniversalConfig},
 };
+use orion_conf::YamlIO;
+use orion_variate::{EnvDict, EnvEvalable};
 use std::env::home_dir;
 
 pub fn build_accessor(dict: &EnvDict) -> UniversalAccessor {
@@ -14,7 +12,7 @@ pub fn build_accessor(dict: &EnvDict) -> UniversalAccessor {
         .map(|x| x.join(NET_ACCS_CTRL_PATH_FILE))
         .filter(|p| p.exists())
     {
-        match NetAccessCtrl::from_yml(&path) {
+        match NetAccessCtrl::load_yaml(&path) {
             Ok(ctrl) => {
                 let ctrl = ctrl.env_eval(dict);
                 return UniversalAccessor::new(UniversalConfig::default().with_ctrl(ctrl));