+++ src/util/git.rs
@@ -1,9 +1,9 @@
+use orion_accessor::addr::Address;
+use orion_accessor::addr::GitRepository;
+use orion_accessor::types::ResourceDownloader;
+use orion_accessor::types::UpdateUnit;
+use orion_accessor::update::DownloadOptions;
 use orion_error::ErrorOwe;
-use orion_variate::addr::Address;
-use orion_variate::addr::GitRepository;
-use orion_variate::types::ResourceDownloader;
-use orion_variate::types::UpdateUnit;
-use orion_variate::update::DownloadOptions;
 use orion_variate::vars::EnvDict;
 
 use crate::ExecResult;