+++ tests/init_cmd_test.rs
@@ -6,9 +6,9 @@ use std::path::PathBuf;
 use galaxy_flow::GxLoader;
 use galaxy_flow::infra::once_init_log;
 use galaxy_flow::util::path::WorkDir;
+use orion_accessor::addr::GitRepository;
 use orion_error::TestAssertWithMsg;
 use orion_infra::path::ensure_path;
-use orion_variate::addr::GitRepository;
 
 // use shells;
 #[ignore]